move debug to new file
