+++ packages/desktop/app/components/Table/Table.js
@@ -267,8 +267,7 @@ const StatusRow = React.memo(({ colSpan, children, textColor }) => (
 class TableComponent extends React.Component {
   getStatusMessage() {
     const { isLoading, errorMessage, data, noDataMessage } = this.props;
-    if (isLoading)
-      return <TranslatedText stringId="general.status.loading" fallback="Loading..." />;
+    if (isLoading) return <TranslatedText stringId="general.table.loading" fallback="Loading..." />;
     if (errorMessage) return errorMessage;
     if (!data.length) return noDataMessage;
     return null;
